requirements.txt
# Siigo API CLI Application - Requirements

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Additional utilities
pathlib2>=2.3.7; python_version < '3.4'